requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.4.1
psycopg2-binary
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0